+++ wxc_sdk/as_api.py
@@ -16635,6 +16635,9 @@ class AsManageNumbersJobsApi(AsApiChild, base='telephony/config/jobs/numbers'):
             * Number Usage Type can be set to `NONE` if carrier has the PSTN service `GEOGRAPHIC_NUMBERS`.
             * Number Usage Type can be set to `SERVICE` if carrier has the PSTN service `SERVICE_NUMBERS`.
 
+        For example, you can initiate a `NUMBER_USAGE_CHANGE` job to change the number type from Standard number to
+        Service number, or the other way around.
+
         :param operation: Indicates the kind of operation to be carried out.
         :type operation: str
         :param number_list: Numbers on which to execute the operation.
@@ -23502,16 +23505,22 @@ class AsTelephonyApi(AsApiChild, base='telephony/config'):
                       state: NumberState = None, details: bool = None, toll_free_numbers: bool = None,
                       restricted_non_geo_numbers: bool = None,
                       included_telephony_type: TelephonyType = None,
+                      service_number: bool = None,
                       org_id: str = None, **params) -> AsyncGenerator[NumberListPhoneNumber, None, None]:
         """
-        Get Phone Numbers for an Organization with given criteria.
+        Get Phone Numbers for an Organization with Given Criterias
 
         List all the phone numbers for the given organization along with the status and owner (if any).
 
-        PSTN phone numbers are associated with a specific location and can be active/inactive and assigned/unassigned.
-        The owner is the person, workspace, or feature to which the number is assigned.
-        Retrieving this list requires a full or read-only administrator auth token with a scope
-        of spark-admin:telephony_config_read.
+        Numbers can be standard, service, or mobile. Both standard and service numbers are PSTN numbers.
+        Service numbers are considered as high-utilization or high-concurrency phone numbers and can be assigned to
+        features like auto-attendants, call queues, and hunt groups.
+        Phone numbers can be linked to a specific location, be active or inactive, and be assigned or unassigned.
+        The owner of a number is the person, workspace, or feature to which the number is assigned.
+        Only a person can own a mobile number.
+
+        Retrieving this list requires a full or read-only administrator or location administrator auth token with a
+        scope of `spark-admin:telephony_config_read`.
 
         :param location_id: Return the list of phone numbers for this location within the given organization.
         :type location_id: str
@@ -23550,6 +23559,8 @@ class AsTelephonyApi(AsApiChild, base='telephony/config'):
             By default if this query parameter is not provided, it will list both PSTN and Mobile Numbers. Possible
             input values are PSTN_NUMBER, MOBILE_NUMBER.
         :type included_telephony_type: TelephonyType
+        :param service_number: Returns the list of service phone numbers.
+        :type service_number: bool
         :param org_id: List numbers for this organization.
         :type org_id: str
         :return: yields :class:`NumberListPhoneNumber` instances
@@ -23579,16 +23590,22 @@ class AsTelephonyApi(AsApiChild, base='telephony/config'):
                       state: NumberState = None, details: bool = None, toll_free_numbers: bool = None,
                       restricted_non_geo_numbers: bool = None,
                       included_telephony_type: TelephonyType = None,
+                      service_number: bool = None,
                       org_id: str = None, **params) -> List[NumberListPhoneNumber]:
         """
-        Get Phone Numbers for an Organization with given criteria.
+        Get Phone Numbers for an Organization with Given Criterias
 
         List all the phone numbers for the given organization along with the status and owner (if any).
 
-        PSTN phone numbers are associated with a specific location and can be active/inactive and assigned/unassigned.
-        The owner is the person, workspace, or feature to which the number is assigned.
-        Retrieving this list requires a full or read-only administrator auth token with a scope
-        of spark-admin:telephony_config_read.
+        Numbers can be standard, service, or mobile. Both standard and service numbers are PSTN numbers.
+        Service numbers are considered as high-utilization or high-concurrency phone numbers and can be assigned to
+        features like auto-attendants, call queues, and hunt groups.
+        Phone numbers can be linked to a specific location, be active or inactive, and be assigned or unassigned.
+        The owner of a number is the person, workspace, or feature to which the number is assigned.
+        Only a person can own a mobile number.
+
+        Retrieving this list requires a full or read-only administrator or location administrator auth token with a
+        scope of `spark-admin:telephony_config_read`.
 
         :param location_id: Return the list of phone numbers for this location within the given organization.
         :type location_id: str
@@ -23627,6 +23644,8 @@ class AsTelephonyApi(AsApiChild, base='telephony/config'):
             By default if this query parameter is not provided, it will list both PSTN and Mobile Numbers. Possible
             input values are PSTN_NUMBER, MOBILE_NUMBER.
         :type included_telephony_type: TelephonyType
+        :param service_number: Returns the list of service phone numbers.
+        :type service_number: bool
         :param org_id: List numbers for this organization.
         :type org_id: str
         :return: yields :class:`NumberListPhoneNumber` instances